robustness(avoid readSocket to block infinitely)
HTTPClient uses readSocket as part of the connection and it is a blocking method. It may happen that the readSocket is stuck and taking a lot of time making the connector executor thread uninterruptable by the Engine: the connector executor timeout will have not effect.

Without any socket timeout specified in the request builder, the JVM will handle the timeout using OS default config which can be really long. This meand that the connector executor pool can be clogged in no time and the only solution is to restart the server which is not robust.